[Java] Comparisons emit calls to Values.compare()
[Java] Comparison to use Value.equals() for equality

To simplify the code generation for Value.compare(), we use .compareTo
which limits the types of arguments we can pass to Value.compare()
(for instance, Lists and Maps don't implement that interface).  This
is fine because P doesn't let programmers test ordering on those types,
just equality.  So, this means we have to split the two up so == and !=
work on all P values.